+++ src/modules/database/types/database-cluster-connection.ts
@@ -1,0 +1,9 @@
+export interface IDatabaseClusterConnection {
+  database: string;
+  host: string;
+  password: string;
+  port: number;
+  ssl: boolean;
+  uri: string;
+  user: string;
+}